OperatorImpl file + ExecutionContext struct
