+++ operators/src/engine/operator.rs
@@ -3,45 +3,11 @@ use super::{
     RasterResultDescriptor, VectorResultDescriptor,
 };
 use crate::util::Result;
-use geoengine_datatypes::collections::VectorDataType;
-use geoengine_datatypes::{projection::ProjectionOption, raster::RasterDataType};
+
 use serde::{Deserialize, Serialize};
 
 /// Common methods for `Operator`s
 pub trait Operator: std::fmt::Debug + Send + Sync + CloneableOperator {}
-
-#[derive(Debug, Serialize, Deserialize, Clone)]
-pub struct OperatorImpl<P> {
-    pub params: P,
-    pub raster_sources: Vec<Box<dyn RasterOperator>>,
-    pub vector_sources: Vec<Box<dyn VectorOperator>>,
-}
-
-impl<P> Operator for OperatorImpl<P> where P: std::fmt::Debug + Send + Sync + Clone + 'static {}
-
-#[derive(Debug, Serialize, Deserialize, PartialEq, Clone)]
-pub struct SourceOperatorImpl<P> {
-    pub params: P,
-}
-
-impl<P> Operator for SourceOperatorImpl<P> where P: std::fmt::Debug + Send + Sync + Clone + 'static {}
-
-/// Common methods for `VectorOperator`s
-#[typetag::serde(tag = "type")]
-pub trait VectorOperator: Operator + CloneableVectorOperator {
-    fn into_initialized_operator(
-        self: Box<Self>,
-        context: ExecutionContext,
-    ) -> Result<Box<dyn InitializedVectorOperator>>;
-
-    /// Wrap a box around a `VectorOperator`
-    fn boxed(self) -> Box<dyn VectorOperator>
-    where
-        Self: Sized + 'static,
-    {
-        Box::new(self)
-    }
-}
 
 /// Common methods for `RasterOperator`s
 #[typetag::serde(tag = "type")]
@@ -60,7 +26,25 @@ pub trait RasterOperator: Operator + CloneableRasterOperator {
     }
 }
 
-pub type ExecutionContext = u32;
+/// Common methods for `VectorOperator`s
+#[typetag::serde(tag = "type")]
+pub trait VectorOperator: Operator + CloneableVectorOperator {
+    fn into_initialized_operator(
+        self: Box<Self>,
+        context: ExecutionContext,
+    ) -> Result<Box<dyn InitializedVectorOperator>>;
+
+    /// Wrap a box around a `VectorOperator`
+    fn boxed(self) -> Box<dyn VectorOperator>
+    where
+        Self: Sized + 'static,
+    {
+        Box::new(self)
+    }
+}
+
+#[derive(Debug, Copy, Clone, PartialEq)]
+pub struct ExecutionContext;
 
 pub trait InitializedOperator {
     fn execution_context(&self) -> &ExecutionContext;
@@ -124,93 +108,6 @@ impl InitializedVectorOperator for Box<dyn InitializedVectorOperator> {
     }
     fn vector_processor(&self) -> Result<TypedVectorQueryProcessor> {
         self.as_ref().vector_processor()
-    }
-}
-
-pub struct InitilaizedOperatorImpl<Parameters, ResultDescriptor> {
-    pub params: Parameters,
-    pub raster_sources: Vec<Box<dyn InitializedRasterOperator>>,
-    pub vector_sources: Vec<Box<dyn InitializedVectorOperator>>,
-    pub context: ExecutionContext,
-    pub result_descriptor: ResultDescriptor,
-}
-
-impl<P, R> InitilaizedOperatorImpl<P, R> {
-    pub fn new(
-        params: P,
-        context: ExecutionContext,
-        result_descriptor: R,
-        raster_sources: Vec<Box<dyn InitializedRasterOperator>>,
-        vector_sources: Vec<Box<dyn InitializedVectorOperator>>,
-    ) -> Self {
-        Self {
-            params,
-            raster_sources,
-            vector_sources,
-            context,
-            result_descriptor,
-        }
-    }
-
-    pub fn create<F>(
-        params: P,
-        context: ExecutionContext,
-        result_descriptor_fn: F,
-        raster_sources_not_init: Vec<Box<dyn RasterOperator>>,
-        vector_sources_not_init: Vec<Box<dyn VectorOperator>>,
-    ) -> Result<Self>
-    where
-        F: Fn(
-            &P,
-            &ExecutionContext,
-            &[Box<dyn InitializedRasterOperator>],
-            &[Box<dyn InitializedVectorOperator>],
-        ) -> Result<R>,
-    {
-        let raster_sources = raster_sources_not_init
-            .into_iter()
-            .map(|o| o.initialized_operator(context))
-            .collect::<Result<Vec<Box<dyn InitializedRasterOperator>>>>()?;
-        let vector_sources = vector_sources_not_init
-            .into_iter()
-            .map(|o| o.into_initialized_operator(context))
-            .collect::<Result<Vec<Box<dyn InitializedVectorOperator>>>>()?;
-        let result_descriptor = result_descriptor_fn(
-            &params,
-            &context,
-            raster_sources.as_slice(),
-            vector_sources.as_slice(),
-        )?;
-
-        Ok(Self::new(
-            params,
-            context,
-            result_descriptor,
-            raster_sources,
-            vector_sources,
-        ))
-    }
-}
-
-impl<P, R> InitializedOperator for InitilaizedOperatorImpl<P, R>
-where
-    P: std::fmt::Debug + Clone,
-    R: std::fmt::Debug + Clone,
-{
-    fn execution_context(&self) -> &ExecutionContext {
-        &self.context
-    }
-    fn raster_sources(&self) -> &[Box<dyn InitializedRasterOperator>] {
-        self.raster_sources.as_slice()
-    }
-    fn vector_sources(&self) -> &[Box<dyn InitializedVectorOperator>] {
-        self.vector_sources.as_slice()
-    }
-    fn raster_sources_mut(&mut self) -> &mut [Box<dyn InitializedRasterOperator>] {
-        self.raster_sources.as_mut_slice()
-    }
-    fn vector_sources_mut(&mut self) -> &mut [Box<dyn InitializedVectorOperator>] {
-        self.vector_sources.as_mut_slice()
     }
 }
